Fix mono and nativeaot call sites, and share identical ListBuilder between coreclr/mono
